requirements.txt
aiofiles==23.2.1
aiosmtplib==2.0.2
alembic==1.12.0
amqp==5.1.1
annotated-types==0.6.0
anyio==3.7.1
apache-libcloud==3.8.0
asyncpg==0.28.0
bcrypt==4.0.1
billiard==4.1.0
blinker==1.6.3
celery==5.3.4
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.4
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.104.0
fastapi-mail==1.4.1
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
flower==2.0.1
greenlet==3.0.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
humanize==4.8.0
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
orjson==3.9.1
packaging==23.2
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
PyJWT==2.8.0
pytest==7.4.2
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
sqlalchemy-file==0.6.0
starlette==0.27.0
tomli==2.0.1
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
ujson==5.8.0
urllib3==2.0.6
uvicorn==0.22.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.8
websockets==11.0.3
